tests/wpt/tests/css/css-writing-modes/unicode-bidi-parsing-001.html

<!DOCTYPE html>
<title>CSS Writing Modes: parsing unicode-bidi: normal, embed, bidi-override</title>
<link rel="author" title="Koji Ishii" href="mailto:kojiishi@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css2/visuren.html#direction">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#unicode-bidi">
<meta name="assert" content="This test asserts the parser and getComputedStyle works correctly for the unicode-bidi: normal, embed, bidi-override.">
<meta name="flags" content="dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<span title="Initial value of unicode-bidi"
  data-expected="normal"></span>

<div style="unicode-bidi: embed; unicode-bidi: normal"
  data-expected="normal"></div>
<div style="unicode-bidi: embed"
  data-expected="embed"></div>
<div style="unicode-bidi: bidi-override"
  data-expected="bidi-override"></div>

<div style="unicode-bidi: embed">
  <span title="unicode-bidi should not inherit"
    data-expected="normal"></span>
</div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).unicodeBidi;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
