LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/unicode-bidi-parsing-001-expected.txt


FAIL Initial value of unicode-bidi assert_equals: expected "normal" but got "isolate"
PASS unicode-bidi: embed; unicode-bidi: normal
PASS unicode-bidi: embed
PASS unicode-bidi: bidi-override
FAIL unicode-bidi should not inherit assert_equals: expected "normal" but got "isolate"
